🐸🦀 Checkpoint ./Handle_Clear.js:2579398/380

+++ Handle_Clear.js
@@ -1,5 +1,6 @@
 // args[1] is the # of messages to delete
 const handleClear = function(msg, args) {
+  if (!msg.client.hasPermission('KICK_MEMBERS')) return msg.channel.sendMessage(`You don't have permission to delete messages you fool`);
   if(!args[1]) return msg.channel.sendMessage('Please specify # of messages to delete');
   msg.channel.bulkDelete(args[1] + 1)   // +1 because this creates a message itself
     .then(() => msg.channel.sendMessage(`Cleared ${args[1]} messages`))
